pyproject.toml
[project]
name = "fastapi-gh-wehbook"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.38.12",
    "docker>=7.1.0",
    "fastapi>=0.115.12",
    "gunicorn>=23.0.0",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "uvicorn>=0.34.2",
]
